+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutClient.java
@@ -11,9 +11,8 @@ import com.paypal.checkout.config.CheckoutConfig;
 import com.paypal.checkout.config.Environment;
 import com.paypal.checkout.config.SettingsConfig;
 import com.paypal.checkout.config.UIConfig;
-
-import com.paypal.pyplcheckout.common.instrumentation.PEnums;
-import com.paypal.pyplcheckout.common.instrumentation.PLog;
+import com.paypal.pyplcheckout.instrumentation.constants.PEnums;
+import com.paypal.pyplcheckout.instrumentation.di.PLog;
 
 import org.json.JSONException;
 import org.json.JSONObject;